// src/actions/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Fri, 14 Jul 2023 17:23:04 GMT
 */

/**
 * @enum
 * @readonly
 * @description EscrwWorlds Action Names
 */
export enum EscrwWorldsActionName {
    Approve = 'approve',
    Cancel = 'cancel',
    Clean = 'clean',
    Disapprove = 'disapprove',
    Dispute = 'dispute',
    Init = 'init',
    Refund = 'refund',
    Transfer = 'transfer',
};